advice-img-databroker/requirements.txt
# GRPC Python setup requirements
coverage>=4.0
cython>=0.29.8
enum34>=1.0.4
six>=1.15.0
# numpy>=1.19.5
wheel>=0.29
googleapis-common-protos>=1.53.0
grpcio>=1.37.0
grpcio-tools>=1.37.0
protobuf>=3.15.8


